fix(eve): harden sandbox credential lifecycle
