+++ packages/eve/src/public/sandbox/vercel-sandbox.ts
@@ -4,7 +4,7 @@ import type {
   NetworkPolicyRule,
   NetworkTransformer,
 } from "#compiled/@vercel/sandbox/index.js";
-import type { ConnectionAuthDefinition, TokenResult } from "#runtime/connections/types.js";
+import type { ConnectionAuthProvider, TokenResult } from "#runtime/connections/types.js";
 
 type VercelCreateOptions = NonNullable<Parameters<typeof Vercel.Sandbox.create>[0]>;
 
@@ -19,42 +19,23 @@ type VercelSandboxAuthorCreateOptions<T> = T extends unknown
   : never;
 
 /**
- * Options accepted by `vercel(opts)`. Forwarded to Vercel
- * Sandbox creation for every fresh sandbox the framework creates
- * (template at prewarm time, session at first-time session-create).
- * Skipped on resume (`Sandbox.get`) since no create happens there.
+ * An eve-managed authenticated Vercel firewall rule.
  *
- * `networkPolicy` is deferred until after framework-owned base setup
- * for fresh templates and template-less sessions, so eve can install
- * required packages before authored bootstrap code runs. Template-backed
- * session creates receive it at creation time because the template
- * already contains the prepared base runtime.
- *
- * Framework-injected fields (`name`, `onResume`, `persistent`, `signal`)
- * are excluded: the framework owns those and overrides any
- * author-supplied values.
- *
- * `source` is honored only on the template create at prewarm time, so
- * an author-supplied snapshot, git revision, or tarball becomes the
- * base layer for the template. Framework setup, bootstrap, and seed
- * files all run on top, and the resulting
- * framework-owned snapshot is what every later session derives from,
- * so `source` is stripped from the session-create path. eve does not
- * detect external snapshot changes; to pick up a rebuilt external
- * snapshot, force a template rebuild (e.g. by changing the sandbox
- * definition so its template key changes).
+ * Interactive providers can initiate authorization when an authored tool's
+ * `execute` first opens the sandbox. Hooks and channel events can reuse an
+ * already-authorized provider but cannot initiate this interactive flow.
  */
-/** An Eve-managed authenticated Vercel firewall rule. */
 export interface VercelSandboxAuthNetworkPolicyRule {
-  readonly auth: ConnectionAuthDefinition;
+  /** Static authorization provider resolved for the active principal. */
+  readonly auth: ConnectionAuthProvider;
   readonly match?: NetworkPolicyMatch;
   readonly transform: (token: TokenResult) => NetworkTransformer[];
 }
 
-/** A native Vercel rule or an Eve-managed authenticated rule. */
+/** A native Vercel rule or an eve-managed authenticated rule. */
 export type VercelSandboxNetworkPolicyRule = NetworkPolicyRule | VercelSandboxAuthNetworkPolicyRule;
 
-/** Route-level Vercel policy shape accepted by Eve. */
+/** Route-level Vercel policy shape accepted by eve. */
 export type VercelSandboxNetworkPolicy =
   | "allow-all"
   | "deny-all"
@@ -69,10 +50,31 @@ export type VercelSandboxNetworkPolicy =
     };
 
 /**
- * Options accepted by `vercel(opts)`.
+ * Options accepted by `vercel(opts)`. Forwarded to Vercel Sandbox creation
+ * for every fresh sandbox the framework creates (template at prewarm time,
+ * session at first-time session-create). Skipped on resume (`Sandbox.get`)
+ * since no create happens there.
+ *
+ * `networkPolicy` is deferred until after framework-owned base setup for
+ * fresh templates and template-less sessions, so eve can install required
+ * packages before authored bootstrap code runs. Template-backed session
+ * creates receive it at creation time because the template already contains
+ * the prepared base runtime.
+ *
+ * Framework-injected fields (`name`, `onResume`, `persistent`, `signal`) are
+ * excluded: the framework owns those and overrides author-supplied values.
+ *
+ * `source` is honored only on the template create at prewarm time, so an
+ * author-supplied snapshot, git revision, or tarball becomes the base layer
+ * for the template. Framework setup, bootstrap, and seed files all run on
+ * top, and the resulting framework-owned snapshot is what every later
+ * session derives from, so `source` is stripped from the session-create path.
+ * eve does not detect external snapshot changes; to pick up a rebuilt
+ * external snapshot, force a template rebuild (e.g. by changing the sandbox
+ * definition so its template key changes).
  *
  * The Vercel SDK create options remain available. Record-form `networkPolicy`
- * rules may attach `auth` for Eve-managed credential resolution.
+ * rules may attach `auth` for eve-managed credential resolution.
  */
 export type VercelSandboxCreateOptions = VercelSandboxCreateOptionsWithAuth<VercelCreateOptions>;
